% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{This function creates a calibration object. A calibration object is used in a model family
to specify how to compare models in a model family to determine the best model.}
\usage{
calibration(score = "accuracy", sampling = "bootstrap", reps = 10)
}
\arguments{
\item{score}{Defines the score to be used for the calibration object. Defaults to 'accuracy'.}
}
\value{
The new calibration object. This is of class "calibration".
}
\description{
This function creates a calibration object. A calibration object is used in a model family
to specify how to compare models in a model family to determine the best model.
}
\examples{
cal <- calibration(score = "accuracy",sampling = "bootstrap",reps = 10)
}
