% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.R
\name{extract_model}
\alias{extract_model}
\title{This function gives the best model from a fitted model family.}
\usage{
extract_model(modfam)
}
\arguments{
\item{modfam}{The fitted model family.}
}
\value{
The best model in the fitted model family.
}
\description{
This function gives the best model from a fitted model family.
}
\examples{
best_model <- extract_model(fitted_model_fam)
}
