% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.model}
\alias{fit.model}
\title{This specifies what to do when an object of class "model" is passed into the fit function.
This will call the fitmodel function on the model with it's corresponding parameters to
fit the model.}
\usage{
\method{fit}{model}(mod, hp_space, hp_strategy, calibration, k = -1,
  metric = "euclidian")
}
\arguments{
\item{mod}{The model to be fit.}

\item{hp_space}{The hyperparameter space. This is unused for objects of class "model".}

\item{hp_strategy}{The hyperparameter strategy. This is unused for objects of class "model".}

\item{calibration}{The calibration. This is unused for objects of class "model".}

\item{k}{The k value for the model. Can also be any generic parameter value. Default is set
to -1 which is calculated to a more sane default when passed into the fitmodel function.}

\item{metric}{The metric to use for fitting the model. Default is set to "eculidian".}
}
\value{
The fitted model for mod. This is of class "model".
}
\description{
This specifies what to do when an object of class "model" is passed into the fit function.
This will call the fitmodel function on the model with it's corresponding parameters to
fit the model.
}
\examples{
fit_mod <- fit(mod = new_knn_model(data=iris, design = Species ~ .), k=3, metric="euclidian")
}
