% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.modelfamily}
\alias{fit.modelfamily}
\title{This specifies what to do when an object of class "modelfamily" is passed into the fit function.
This will call the fitmodelfamily function on the modelfamily with it's corresponding parameters to
fit the model family.}
\usage{
\method{fit}{modelfamily}(mod, hp_space, hp_strategy, calibration, k, metric)
}
\arguments{
\item{mod}{The model family to be fit.}

\item{hp_space}{The hyperparameter space. This specifies the hyperparameters in the model
family.}

\item{hp_strategy}{The hyperparameter strategy. This specifies how the hyperparameters
are selected when fitting the model family.}

\item{calibration}{The calibration. This specifies how each model will be assessed when
fitting the model family.}

\item{k}{The k value. This is unused for objects of class "modelfamily".}

\item{metric}{The metric. This is unused for objects of class "modelfamily".}
}
\value{
The fitted modelfamily for mod. This is of class "modelfamily".
}
\description{
This specifies what to do when an object of class "modelfamily" is passed into the fit function.
This will call the fitmodelfamily function on the modelfamily with it's corresponding parameters to
fit the model family.
}
\examples{
fitted_model_fam <- fit(mod = new_knn_modelfamily(data = iris,design = Species ~ .),
hp_space = hp_space(vals = c(5,3,1), metric = c("euclidian")),
hp_strategy = "grid_search",
calibration = calibration(score = "accuracy",sampling = "bootstrap",reps = 10))
}
