% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{This is a generic function for fit. It can be called on objects of any class.
For objects that are not models or model family's it simply prints a message
stating that you are "in fit.default".}
\usage{
fit(mod, hp_space, hp_strategy, calibration, k, metric)
}
\arguments{
\item{mod}{The model or model family object. This will likely be empty in the default case.}

\item{hp_space}{The hyperparameter space. This will likely be empty in the default case.}

\item{hp_strategy}{The hyperparameter strategy. This will likely be empty in the default case.}

\item{calibration}{The calibration. This will likely be empty in the default case.}

\item{k}{The k value. This will likely be empty in the default case.}

\item{metric}{The metric. This will likely be empty in the default case.}
}
\description{
This is a generic function for fit. It can be called on objects of any class.
For objects that are not models or model family's it simply prints a message
stating that you are "in fit.default".
}
\examples{
fit()
}
