% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fitmodel.knn}
\alias{fitmodel.knn}
\title{This function specifically handles fitting knn models. This handles
everything that needs to be done when fitting a knn model.}
\usage{
\method{fitmodel}{knn}(mod, k = -1, metric = "euclidian")
}
\arguments{
\item{mod}{The knn model.}

\item{k}{The k value for the model. The default in this case in -1 which in the function
is changed to the calculation of the square root of the number of rows of data.}

\item{metric}{The metric. The default is set to "euclidian".}
}
\value{
The fitted knn model. This is an object of classes "knn" and "model".
}
\description{
This function specifically handles fitting knn models. This handles
everything that needs to be done when fitting a knn model.
}
\examples{
fitmodel(mod = new_knn_model(data=iris, design = Species ~ .), k=3, metric="euclidian")
}
