% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fitmodel}
\alias{fitmodel}
\title{This is a generic function that specifically handles fitting for models. This handles
everything that needs to be done when fitting a model. For objects that do not have a
defined class that specifies the kind of model this simply prints "in fitmodel.default".}
\usage{
fitmodel(mod, k, metric)
}
\arguments{
\item{mod}{The model object. This will likely be empty in the default case.}

\item{k}{The k value. This will likely be empty in the default case.}

\item{metric}{The metric. This will likely be empty in the default case.}
}
\description{
This is a generic function that specifically handles fitting for models. This handles
everything that needs to be done when fitting a model. For objects that do not have a
defined class that specifies the kind of model this simply prints "in fitmodel.default".
}
\examples{
fitmodel()
}
