% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_modelfamily.R
\name{fitmodelfamily}
\alias{fitmodelfamily}
\title{This is a generic function that specifically handles fitting for model families. This handles
everything that needs to be done when fitting a model family. For objects that are not of a
defined class for model families this simply prints "in fitmodel.default".}
\usage{
fitmodelfamily(mod, hp_space, hp_strategy, calibration)
}
\arguments{
\item{mod}{The modelfamily object. This will likely be empty in the default case.}

\item{hp_space}{The hyperparameter space. This specifies the hyperparameters in the model
family. This will likely be empty in the default case.}

\item{hp_strategy}{The hyperparameter strategy. This specifies how the hyperparameters
are selected when fitting the model family. This will likely be empty in the default case.}

\item{calibration}{The calibration. This specifies how each model will be assessed when
fitting the model family. This will likely be empty in the default case.}
}
\description{
This is a generic function that specifically handles fitting for model families. This handles
everything that needs to be done when fitting a model family. For objects that are not of a
defined class for model families this simply prints "in fitmodel.default".
}
\examples{
fitmodelfamily()
}
