% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_space.R
\name{hp_space}
\alias{hp_space}
\title{This function creates a hyperparameter space object. A hyperparameter space object is
used in model families to specify the hyperparameters in the model family.}
\usage{
hp_space(vals = NULL, metric = NULL)
}
\arguments{
\item{vals}{Defines the hyperparameter values.}
}
\value{
The new hyperparameter space object. This is of class "hp_space".
}
\description{
This function creates a hyperparameter space object. A hyperparameter space object is
used in model families to specify the hyperparameters in the model family.
}
\examples{
cal <- calibration(score = "accuracy",sampling = "bootstrap",reps = 10)
}
