% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_knn_model.R
\name{new_knn_model}
\alias{new_knn_model}
\title{This function creates a knn model object.}
\usage{
new_knn_model(data = NULL, design = NULL)
}
\arguments{
\item{data}{Defines the data to be used for the knn model object.}
}
\value{
The new knn model object. This is of classes "knn" and "model". This model object
has a boolean indicating it hasn't been fit, a default k value of -1, a default metric of
"euclidian, initialized empty values for predicted, truth, and train_data, and the data and
design fields set to be equal to the parameters passed into this function.
}
\description{
This function creates a knn model object.
}
\examples{
unfit_mod <- new_knn_model(data=iris, design = Species ~ .)
}
