% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_knn_modelfamily.R
\name{new_knn_modelfamily}
\alias{new_knn_modelfamily}
\title{This function creates a knn model family object.}
\usage{
new_knn_modelfamily(data = NULL, design = NULL)
}
\arguments{
\item{data}{Defines the data to be used for the knn model family object.}
}
\value{
The new knn model famil object. This is of classes "knn" and "modelfamily".
 This model object has a boolean indicating it hasn't been fit, initialized empty value
 for best_model, and the data and design fields set to be equal to the parameters
 passed into this function.
}
\description{
This function creates a knn model family object.
}
\examples{
model_fam <- new_knn_modelfamily(data = iris,design = Species ~ .)
}
