% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.model}
\alias{predict.model}
\title{This specifies what to do when predict is called on an object of class "model". This will
call the function predictmodel on the model with it's corresponding parameters to
predict the values for the new_data using the model.}
\usage{
\method{predict}{model}(mod, new_data)
}
\arguments{
\item{mod}{The model object. This model must be fitted.}

\item{newdata}{The new data you want to predict values for.}
}
\value{
The predicted values for the new_data passed in.
}
\description{
This specifies what to do when predict is called on an object of class "model". This will
call the function predictmodel on the model with it's corresponding parameters to
predict the values for the new_data using the model.
}
\examples{
mod_predicts <- predict(fit_mod, new_data)
}
