% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.modelfamily}
\alias{predict.modelfamily}
\title{This specifies what to do when predict is called on an object of class "modelfamily". This will
call the function predictmodelfamily on the fitted model family with it's corresponding parameters to
predict the values for the new data using the best model in the fitted model family.}
\usage{
\method{predict}{modelfamily}(mod, new_data)
}
\arguments{
\item{mod}{The model famil object. This model family must be fitted.}

\item{newdata}{The new data you want to predict values for.}
}
\value{
The predicted values for the new_data passed in.
}
\description{
This specifies what to do when predict is called on an object of class "modelfamily". This will
call the function predictmodelfamily on the fitted model family with it's corresponding parameters to
predict the values for the new data using the best model in the fitted model family.
}
\examples{
mod_family_predicts <- predict(fitted_model_fam, new_data)
}
