% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{This is a generic function for predict. It can be called on objects of any class.
For objects that are not models or model family's it simply prints a message
stating that you are "in predict.default".}
\usage{
predict(mod, new_data)
}
\arguments{
\item{mod}{The model or model family object. This will likely be empty in the default case.}

\item{newdata}{The new data you want to predict values for. This will likely be empty in
the default case.}
}
\description{
This is a generic function for predict. It can be called on objects of any class.
For objects that are not models or model family's it simply prints a message
stating that you are "in predict.default".
}
\examples{
predict()
}
