% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predictmodel.knn}
\alias{predictmodel.knn}
\title{This function specifically handles predicting knn models. This handles
everything that needs to be done when predicting a knn model.}
\usage{
\method{predictmodel}{knn}(mod, new_data)
}
\arguments{
\item{mod}{The model object.}

\item{newdata}{The new data you want to predict values for.}
}
\value{
The predicted values for the new_data passed in.
}
\description{
This function specifically handles predicting knn models. This handles
everything that needs to be done when predicting a knn model.
}
\examples{
predictmodel(fit_mod, new_data)
}
