% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predictmodel}
\alias{predictmodel}
\title{This is a generic function that specifically handles predicting for models. This handles
everything that needs to be done when predicting a model. For objects that do not have a
defined class that specifies the kind of model this simply prints "in predictmodel.default".}
\usage{
predictmodel(mod, new_data)
}
\arguments{
\item{mod}{The model object. This will likely be empty in the default case.}

\item{newdata}{The new data you want to predict values for. This will likely be empty in
the default case.}
}
\description{
This is a generic function that specifically handles predicting for models. This handles
everything that needs to be done when predicting a model. For objects that do not have a
defined class that specifies the kind of model this simply prints "in predictmodel.default".
}
\examples{
predictmodel()
}
