% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_modelfamily.R
\name{predictmodelfamily.knn}
\alias{predictmodelfamily.knn}
\title{This function specifically handles predicting knn model families. This handles
everything that needs to be done when predicting a knn model family.}
\usage{
\method{predictmodelfamily}{knn}(mod, new_data)
}
\arguments{
\item{mod}{The model family object.}

\item{newdata}{The new data you want to predict values for.}
}
\value{
The predicted values for the new_data passed in.
}
\description{
This function specifically handles predicting knn model families. This handles
everything that needs to be done when predicting a knn model family.
}
\examples{
predictmodelfamily(fitted_model_fam, new_data)
}
