% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_modelfamily.R
\name{predictmodelfamily}
\alias{predictmodelfamily}
\title{This is a generic function that specifically handles predicting for model families. This handles
everything that needs to be done when predicting a model family. For objects that do not have a
defined class that specifies the kind of model family this simply prints "in predictmodelfamily.default".}
\usage{
predictmodelfamily(mod, new_data)
}
\arguments{
\item{mod}{The model family object. This will likely be empty in the default case.}

\item{newdata}{The new data you want to predict values for. This will likely be empty in
the default case.}
}
\description{
This is a generic function that specifically handles predicting for model families. This handles
everything that needs to be done when predicting a model family. For objects that do not have a
defined class that specifies the kind of model family this simply prints "in predictmodelfamily.default".
}
\examples{
predictmodelfamily()
}
